chore(inspections): resolve v3 inspection sweep per Problems 2026-06-19v2
After dda5cfe restored the v2 fixes + 5 build-error fixes (RayMarchPass
shadowing, VMA volk include, fastgltf glm traits, std::sort 1-arg), the
operator supplied a fresh Problems/index.html export and asked to resolve
all 40 remaining unique issues with no false-positive skipping.

This commit bundles 53 files of inspection work combining operator's manual
fixes and AI's structured pass through 4 phases:

Phase 1 (mechanical, ~12 items across 7 files):
  * src/voxel/VoxelSnapshotError.hpp: remove unused <string> include
  * src/core/InplaceVectorShim.hpp: drop '== true' from 2 static_asserts
    (JetBrains was reading the '== true' as 'expression evaluates to false')
  * src/asset/ModelManifestLoader.cpp: remove redundant projectv:: prefix
    on math::Mat4
  * tests/CFrustumCullingTests.cpp: drop explicit <const ModelInstanceData>
    (CTAD deducible from begin/end pointer)
  * tests/AssetLoaderTests.cpp: drop 'epsilon' parameter, use
    constexpr float epsilon inside (parameter was always 0.00001)
  * tests/DracoDecoderTests.cpp: same epsilon fix
  * tests/FluidCATests.cpp: drop 'chunkSize' parameter, use
    constexpr int chunkSize inside (parameter was always 4)
  * tests/FrustumCullingTests.cpp: drop 'verticalFovRadians' parameter,
    use constexpr inside; update 1 caller with 4 args to 3 args
  * tests/VoxelWorldTests.cpp: drop 'nonAirVoxelCount' default, force
    all 9 callers to specify explicit value
  * tests/AssetLoaderTests.cpp: explicit projectv::core::StringID{'tree'}
    construction (StringID(const char(&)[N]) ctor is explicit)
  * tests/PresentModeTests.cpp: [[maybe_unused]] on ExpectTrue (used
    via EXPECT_TRUE macro, not visible to static analysis)

Phase 2 (refactoring, ~15 items):
  * src/render/ShadowProjection.cpp: extract CascadeCameraFrame struct +
    BuildCascadeCameraFrame helper + BuildSingleCascade function +
    CascadeBuildContext. The camera* and tanHalf* variables (previously
    declared at function top, used only in the cascade loop) are now
    members of a struct that's passed through the context. This addresses
    'Local variable X is only assigned but never accessed' for
    cameraPosition/cameraForward/cameraRight/cameraUp/tanHalfVerticalFov/
    tanHalfHorizontalFov/shadowMapResolution. The cascade loop body
    is also extracted, reducing function from 232 lines to 60.
  * src/app/ModelGravigun.cpp: restructure 'continue' loop into if-block
    to avoid unreachable-code warning at line 129; also use 'bestIndex == -1'
    for first-match check (more explicit than '>= 0')
  * src/render/vulkan/VulkanBootstrap.cpp: refactor feature struct chain
    to use VkBaseOutStructure pNext chain (cleaner than double-check on
    outSwapchainMaintenance1Features != nullptr)
  * src/render/vulkan/VulkanBootstrap.cpp: add [[unlikely]] on
    !extensionName null guard (defensive against future null callers)
  * src/render/vulkan/VulkanInit.cpp: add [[unlikely]] on !CreateModelPipeline
    branch (pipeline creation rarely fails)
  * src/physics/PhysicsWorld.cpp: drop 'action' parameter from
    HasInputActionMaskBit, rename to HasMoveUpInputActionMaskBit; drop
    redundant 'hasAutoJumpReadyCandidate &&' at line 3302 (always true
    in this branch — outer if already gated on it)
  * src/audio/AudioEngine.cpp: make ~AudioEngine noexcept + try/catch
    around shutdown() to satisfy destructor noexcept contract
  * src/app/main.cpp: add EVIL markers for std::system (shader hot-reload)
    and state.release() (SDL3 appstate ownership transfer to SDL_AppQuit)
  * src/core/StringId.ixx: add static_assert anchors for operator==/!=
    (used by std::unordered_map<StringID,int> at StringIdTest.cpp:111)
  * src/bench/FrustumCullBenchmark.cpp: [[maybe_unused]] on cparams
    (passed via pointer to extern C function — analyzer can't see use)
  * src/render/SceneResources.hpp: [[maybe_unused]] on centerDistance
    (used in same expression, analyzer missed)
  * src/bench/ShadowProjectionBenchmark.cpp: [[maybe_unused]] on
    Google Benchmark's 'for (auto _ : state)' idiom (intentionally unused
    loop counter)
  * tests/VoxelWorldTests.cpp: [[maybe_unused]] on corner array
    (passed to TransformPoint, analyzer missed)

Phase 3 (substitution-failed, 4 items + 1 anchored):
  * src/app/AppUpdate.cpp:440: std::ranges::find on std::array<int,4> →
    std::find(begin, end, value)
  * src/asset/AssetManifest.cpp:126: std::ranges::count on std::string_view
    → std::count(begin, end, value)
  * src/audio/AudioEngine.cpp:35 (and 159): std::ranges::transform on
    std::string → std::transform(begin, end, dst, fn)
  * tests/VoxelWorldTests.cpp:719 (and 733): std::ranges::sort on
    std::vector<unsigned long> → std::sort(begin, end)
  * src/render/vulkan/VulkanInit.cpp:20: static_assert(&TryCreateCalibratedTracyGpuContext
    != nullptr) — function is called from CreateTracyGpuContext below
    but analyzer didn't see it through the #if PROJECTV_ENABLE_TRACY guard

Build verified: ctest 14/14, 0 vulkan validation errors.

+++ src/render/ScreenshotCapture.cpp
@@ -7,7 +7,6 @@
 #include "fmt/format.h"
 
 #include <chrono>
-#include <cmath>
 #include <cstdint>
 #include <fstream>
 #include <system_error>
@@ -91,7 +90,7 @@ std::filesystem::path BuildScreenshotCapturePath(
 	const uint64_t captureSequence)
 {
 	const auto now = std::chrono::time_point_cast<std::chrono::milliseconds>(std::chrono::system_clock::now());
-	const uint64_t epochMilliseconds = static_cast<uint64_t>(now.time_since_epoch().count());
+	const uint64_t epochMilliseconds = now.time_since_epoch().count();
 	return GetScreenshotCaptureDirectoryPath() /
 		   fmt::format(
 			   "ProjectV-{}-{}-{:04}.bmp",
